% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_fname.R
\name{get_fname}
\alias{get_fname}
\title{Get the filename of saved data}
\usage{
get_fname(what, baseDir, iter = NULL, modRow = NULL, subfolder = NULL,
  file_suffix, dirOnly = FALSE)
}
\arguments{
\item{what}{Filename of what kind of data.}

\item{baseDir}{the results' base directory}

\item{iter}{the iteration}

\item{modRow}{the model id/row/parameter setting}
}
\description{
\code{\link{iterativeOcc}} and other functions store intermediate and final results to disc. With this function the filename of specific data packages can be generated.
}
\details{
The following values are possible:
\itemize{
\item ini}
}

