% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iterativeOcc.R
\name{iterativeOcc}
\alias{iterativeOcc}
\title{Train a iterative one-class classifier.}
\usage{
iterativeOcc(P, U, settings = NULL, PN = NULL, overwrite = FALSE,
  verbosity = 3, smallIniGrid = TRUE, tuneGrid = expand.grid(sigma = 1,
  cNeg = 2^seq(-10, 15, 3), cMultiplier = 2^seq(2, 11, 2)), useSigest = TRUE,
  cvPu = TRUE, ...)
}
\arguments{
\item{P}{a data frame with the positive training samples.}

\item{U}{an object of class \code{\link{rasterTiled}} image data to be classified.}

\item{settings}{the settings list}

\item{PN}{a data frame used as independent test set. the first column must be the response variable with 1 being the positive and -1 the negative samples.}

\item{overwrite}{should already computed elements be recalculated}

\item{...}{other arguments that can be passed to \code{\link{trainOcc}}}
}
\description{
...
}

