% Generated by roxygen2 (4.0.1): do not edit by hand
\name{iterativeOcc}
\alias{iterativeOcc}
\title{Train a iterative one-class classifier.}
\usage{
iterativeOcc(train_pos, un, iter_max = 10, n_train_un = nrow(train_pos) * 2,
  k = 10, indep_un = 0.5, folder_out = "resultsOfIterativeOcc",
  test_set = NULL, seed = NULL)
}
\arguments{
\item{train_pos}{a data frame with the positive training samples.}

\item{un}{an object of class \code{\link{rasterTiled}} image data to be classified.}

\item{iter_max}{maximum number of iterations.}

\item{n_train_un}{the number of unlabeled samples to be used for training and validation.}

\item{k}{the number of folds used for resampling.}

\item{indep_un}{the fraction of unlabeled samples used for validation.}

\item{folder_out}{a folder where the results are written to.}

\item{test_set}{a data frame used as independent test set. the first column must be the response variable with 1 being the positive and -1 the negative samples.}

\item{seed}{a seed point to be set for sampling the unlabeled data and the}
}
\description{
...
}

