% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_iocc}
\alias{plot_iocc}
\title{Create different plots of the iterOcc outcome.}
\usage{
plot_iocc(iocc, iter = NULL, what = "hist", modRow = NULL,
  outfile = NULL)
}
\arguments{
\item{iocc}{The base directory or an iocc object.}

\item{iter}{The iteration}

\item{what}{The type of plot. 'grid', 'hist', 'featurespace',
'test', 'n_unlabeled', or 'time'}
}
\description{
Create different plots of the iterOcc outcome.
}

